Finish the move from the old preload system to modules
The old preload system only had one configuration that it loaded up. The new module 
system now support loading modules specified via a slice. This will also allow 
anyone that would want extend PISC with their own words to do so by creating a 
module. This was partially in preparation for the basic sandboxing of chatbots.