WYSIWYG feedback as commands are typed
Unlike most such implementations around, this one minimally switches to
raw mode. In particular, ctrl-c is still handled by the terminal, and the
terminal is still responsible for echoing characters to screen. That echoing
may be a bad idea; see the backspace handling.

Manual testing:
- type '1 2 + 3 *' with the session starting at the bottom of screen
    (ie. needing to scroll)
- type '1 2 + 3 *' with the session not needing to scroll
- backspace over the command

Not yet working:
- multi-word commands like 'swap'. WYSIWYG state is messed up.
- typing '['. Causes segfault and I don't understand why.

Drops readline.